Add files via upload
Chat Tab Functions
Able to use the in game function to set tabs what types of chats, go to a specific tab.

Reworked the group function, should detect in group, put you in the chat tab for group, and switch channel to group.  When group ends it should take back to previous tab, and change group tab chat channel to whatever it was before.

Default channels set all tabs to specific channels and will over write last channel used.  

working on locking the default channels to the tab, can send a chat to another channel but will revert back.